# src/main/resources/static/rest-api-doc.yaml
openapi: 3.0.1
info:
  title: Gwas Catalog Rest API V2
  description:  <p><img width="10%" alt="gwas_icon" src="../GWAS_Catalog_circle_178x178.png"></p>
                <p>Welcome to the <a rel="noopener noreferrer" target="_blank" href="https://www.ebi.ac.uk/gwas">GWAS Catalog</a> Restful API.</p>
                <p>The GWAS RESTful API provides programmatic access to the GWAS Catalog, enables integration of the GWAS Catalog data into custom applications and services, allowing external bioinformatics developers to leverage the curated GWAS datasets, and create more extended solutions for the scientific community.</p>
                <p>GWAS RESTful API V2 has been released with various enhancements & improvements over GWAS RESTful API V1. V1 will be deprecated no later than May 2026.<p>
                <p>Note that this API allows access to the literature-curated top associations & metadata (the same data that is available via the <a rel="noopener noreferrer" target="_blank" href="https://www.ebi.ac.uk/gwas">GWAS Catalog Website</a>. A second API enabling access to full genome-side summary statistics is under development.</p>
  version: "2.0"
servers:
- url: https://wwwdev.ebi.ac.uk/gwas/beta/rest/api
  #url: http://localhost:8087/gwas/rest/api
  description: Generated server url
...
paths:
  /v2/unpublished-studies:
    get:
      tags:
      - unpublished-studies
      operationId: getUnpublishedStudies
      summary: A GET request will list all of the Catalog unpublished studies
      parameters:
      - name: disease_trait
        in: query
        description: |-
          Free text description of the trait investigated in this study

          <i>Example</i> : Early-onset Parkinson's disease
        required: false
        schema:
          type: string
      - name: accession_id
        in: query
        description: |-
          The study’s GWAS Catalog accession ID

          <i>Example</i> : GCST90623790
        required: false
        schema:
          type: string
      - name: title
        in: query
        description: |-
          Title of manuscript the GWAS is included in

          <i>Example</i> : Genome-wide association study
        required: false
        schema:
          type: string
      - name: first_author
        in: query
        description: |-
          Last name and initials of first author

          <i>Example</i> : Doe John
        required: false
        schema:
          type: string
      - name: cohort
        in: query
        description: |-
          Discovery stage cohorts used in this study

          <i>Example</i> : BioImage
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: Fields to sort by
        required: false
        schema:
          type: string
          enum:
            - study_accession
            - variant_count
      - name: direction
        in: query
        description: "Sort direction : asc or desc"
        required: false
        schema:
          type: string
          enum:
            - asc
            - desc
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      responses:
        "500":
            description: Internal Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page-unpublished-studies'
...
  /v2/single-nucleotide-polymorphisms:
    get:
      tags:
      - single-nucleotide-polymorphisms
      operationId: getSnps
      summary: A GET request to retrieve all snps in Catalog
      parameters:
      - name: rs_id
        in: query
        description: "SNP rsId; if a haplotype it may include more than one rs\
          \ number (multiple SNPs comprising the haplotype)\n\n<i>Example</i> : rs3093017 "
        required: false
        schema:
          type: string
      - name: bp_location
        in: query
        description: |-
          search for SNPs via their base pair location

          <i>Example</i> : 95000000
        required: false
        schema:
          type: integer
          format: int32
      - name: bp_start
        in: query
        description: |-
          Search for SNPs on a certain chromosome and within a given base pair location range

          <i>Example</i> : 95000000 (To be used in combination with bp_start,bp_end & chromosome)
        required: false
        schema:
          type: integer
          format: int64
      - name: bp_end
        in: query
        description: |-
          Search for SNPs on a certain chromosome and within a given base pair location range

          <i>Example</i> : 95000000 (To be used in combination with bp_start,bp_end & chromosome)
        required: false
        schema:
          type: integer
          format: int64
      - name: pubmed_id
        in: query
        description: |-
          pubmedId of the publication

          <i>Example</i> : 35241825
        required: false
        schema:
          type: string
      - name: chromosome
        in: query
        description: |-
          Search for SNPs on a certain chromosome

          <i>Example</i> : 13 (To be used in combination with bp_start,bp_end & chromosome)
        required: false
        schema:
          type: string
      - name: gene
        in: query
        description: |-
          Search for SNPs via a gene using parameter geneName

          <i>Example</i> : HBS1L
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: Fields to sort by
        required: false
        schema:
          type: string
          enum:
            - location
            - rs_id
      - name: direction
        in: query
        description: "Sort direction : asc or desc"
        required: false
        schema:
          type: string
          enum:
            - asc
            - desc
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page-single-nucleotide-polymorphisms'
  /v2/single-nucleotide-polymorphisms/{rs_id}:
    get:
      tags:
      - single-nucleotide-polymorphisms
      operationId: getSingleNucleotidePolymorphism
      summary: A GET request to retrieve single snp
      parameters:
      - name: rs_id
        in: path
        description: "Strongest SNP; if a haplotype it may include more than one rs\
          \ number (multiple SNPs comprising the haplotype) <br/> <br/><i> Example\
          \ </i> : rs3093017 "
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/single-nucleotide-polymorphisms'
...
  /v2/body-of-works:
    get:
      tags:
      - body-of-works
      operationId: getBodiesOfWork
      summary: A GET request to retrieve all body of works in Catalog
      parameters:
      - name: title
        in: query
        description: "Title of manuscript the GWAS is included in <br/> <br/><i> Example\
          \ </i> : Genome-wide association study"
        required: false
        schema:
          type: string
      - name: first_author
        in: query
        description: "Last name and initials of first author <br/> <br/><i> Example\
          \ </i> : Doe John"
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - "id,DESC"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page-body-of-works'
  /v2/body-of-works/{bow_id}:
    get:
      tags:
      - body-of-works
      operationId: getBodyOfWork
      summary: A GET request to retrieve a single body of work
      parameters:
      - name: bow_id
        in: path
        description: "primary identifier of Body of work table <br/> <br/><i> Example\
          \ </i> : 169062064"
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/body-of-works'
  /v2/body-of-works/{bow_id}/unpublished-studies:
    get:
      tags:
      - body-of-works
      operationId: getUnpublishedStudies_1
      summary: A GET request to retrieve all unpublished studies for a body of work in Catalog
      parameters:
      - name: bow_id
        in: path
        description: "primary identifier of Body of work table <br/> <br/><i> Example\
          \ </i> : 169062064"
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
          default:
            - "id,DESC"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page-unpublished-studies'
...
components:
  schemas:
    link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
        hreflang:
          type: string
        media:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
    page:
      type: object
      properties:
        size:
          type: integer
          format: int64
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
    errors:
      type: object
      properties:
        error_code:
          type: integer
          format: int32
          description: Error code
          example: "ORA-500001"
        error:
          type: string
          description: Error header
          example: "SQL Exception"
        error_message:
          type: string
          description: Error description
        service_path:
          type: string
          description: Endpoint details
          example: "/studies"
        timestamp:
          type: string
          description: Error timestamp
          example: 2021-10-19T22:22:55.015+0000
        report:
          type: object
          additionalProperties:
            type: string
            description: Addition attributes for the error
...
    body-of-works:
      type: object
      properties:
        gcp_id:
          type: string
        title:
          type: string
          description: Title of manuscript the GWAS is included in
          example: Genome-wide association study on susceptibility genes associated with yang-deficiency constitution
        first_author:
          type: string
          description: first author of the publication
          example: John Doe
        publication_date:
          type: string
          format: date-time
          description:  Publication date (online (epub) date if available)
          example: 2022-03-03
        doi:
          type: string
          description: doi for preprint
          example: https://www.medrxiv.org/content/10.1101/2020.06.17.20134015v2
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    page-body-of-works:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
        content:
          type: array
          items:
            $ref: '#/components/schemas/body-of-works'
        page:
          $ref: '#/components/schemas/page'
...
    locations:
      type: object
      properties:
        chromosome_name:
          type: string
          description: Chromosome number associated with rs number
          example: 13
        chromosome_position:
          type: integer
          format: int32
          description:  Chromosomal position associated with rs number
          example: 40983974
        region:
          $ref: '#/components/schemas/regions'
          description: Cytogenetic region associated with rs number
...
    regions:
      type: object
      properties:
        name:
          type: string
          description: Cytogenetic region associated with rs number
          example: 2q37.2
...
    single-nucleotide-polymorphisms:
      type: object
      properties:
        rs_id:
          type: string
          description: The SNP’s rs Id
          example: rs3093017
        merged:
          type: integer
          format: int64
          description: Whether this SNP has been merged with another SNP in a newer genome build
          example: 0
        functional_class:
          type: string
          description: The SNP’s functional class
          example: intron_variant
        current_snp:
          type: string
          description: SNP the present SNP was merged with as part of a more recent genome build
          example: rs3093017
        last_update_date:
          type: string
          format: date-time
          description: The last date this SNP’s mapping information was updated
          example: 2021-10-19T22:22:55.015+0000
        locations:
          type: array
          items:
            $ref: '#/components/schemas/locations'
          description: The SNP’s genomic locations
        maf:
          type: number
          format: double
          description: Minor allele frequency in the 1000 Genomes Phase 3 combined population
          example: 0.39011
        minor_allele:
          type: string
          description: Second most frequent allele in the 1000 Genomes Phase 3 combined population
          example: C
        alleles:
          type: string
          description: Reference allele / alternative allele(s)
          example: C/A/G/T(forward strand)
        most_severe_consequence:
          type: string
          description: Second most frequent allele in the 1000 Genomes Phase 3 combined population
          example: Intron variant
        mapped_genes:
          type: array
          items:
            type: string
            description: Gene(s) overlapping the variant. If a variant is intergenic, the closest 5' and 3' genes are listed
            example: CCR6,HBS1L
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    page-single-nucleotide-polymorphisms:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
        content:
          type: array
          items:
            $ref: '#/components/schemas/single-nucleotide-polymorphisms'
        page:
          $ref: '#/components/schemas/page'
...
    unpublished-studies:
      type: object
      properties:
        genotyping_technology:
          type: string
          description: The method used to genotype variants in the discovery stage
          example: Genome-wide genotyping array
        study_accession:
          type: string
          description: The study’s GWAS Catalog accession ID
          example: GCST90013684
        array_manufacturer:
          type: string
          description: Genotyping platform manufacturer
          example: Affymetrix
        imputation:
          type: boolean
          description: Whether SNPs were imputed
          example: true
        variant_count:
          type: integer
          format: int32
          description: The number of variants analysed in the discovery stage
          example: -1
        summary_statistics_assembly:
          type: string
          description: Genome assembly for the summary statistics
          example: GRCh38
        agreed_to_cc0:
          type: boolean
          description: Agree to CC0 license or not
          example: true
        full_summary_stats:
          type: string
          description: Summary stats FTP file path
          example: http://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90566001-GCST90567000/GCST90566393
        study_tag:
          type: string
          description: A unique free-text label for each genome-wide association study in the publication
          example: white
        array_information:
          type: string
          description: Additional information about the genotyping array
          example: immunochip
        sample_description:
          type: string
          description: Additional  relevant information required for the interpretation of results
          example: sex (males/females), age (adults/children)
        statistical_model:
          type: string
          description: Details of the statistical model used to determine association significance
          example: PrecisionLife combinatorial SNP analysis
        study_description:
          type: string
          description: Additional information about the study
        disease_trait:
          type: string
          description: The trait under investigation
          example: Coronary artery disease
        efo_trait:
          type: string
          description: EFOs that best matching the trait
          example: EFO_0001645
        background_trait:
          type: string
          description: Any background trait(s) shared by all individuals in the GWAS (e.g. in both cases and controls)
          example: Nicotine dependence
        background_efo_trait:
          type: string
          description: EFOs that best match entered background trait
          example: EFO_0003768
        cohort:
          type: string
          description: Cohort(s) represented in the discovery sample
          example: MVP|UKB|CARDIoGRAMplusC4D
        cohort_id:
          type: string
          description: Cohort specific identifier(s) issued to this research study
          example: UK Biobank ID
        title:
          type: string
          description: Title of manuscript the GWAS is included in
          example: Genome-wide association study on susceptibility genes associated with yang-deficiency constitution
        first_Author:
          type: string
          description: Last name and initials of first author
          example: John Doe
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    page-unpublished-studies:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
        content:
          type: array
          items:
            $ref: '#/components/schemas/unpublished-studies'
        page:
          $ref: '#/components/schemas/page'